Moved common stuff out of fieldset and into lib ready for form
