+++ src/form.js
@@ -273,8 +273,7 @@ export default React.createClass({
     render() {
         // Define our helpers for cloneing our children
         let childNames = [];
-        const clonePred = child => child.props && child.props.name;
-        const cloneProps = child => {            
+        const cloneProps = child => {
             warning(!child.ref, `Attempting to attach ref "${child.ref}" to "${child.props.name}" will be bad for your health`);
             warning(childNames.indexOf(child.props.name) === -1, `Duplicate name "${child.props.name}" found. Duplicate fields will be ignored`);
             childNames = childNames.concat(child.props.name);
@@ -290,7 +289,7 @@ export default React.createClass({
 
         const errorRule = errorsRule(this.state);
         const someRule = {
-            predicate: clonePred,
+            predicate: child => child.props && child.props.name, //same as fieldset isFormable
             clone: cloneProps
         };
